test: Move label layout validator into fixtures
